requirements.txt
ansi==0.1.3
asn1crypto==0.24.0
astroid==1.6.1
attrs==17.4.0
beautifulsoup4==4.6.0
bottle==0.12.13
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
colorlog==3.1.0
cryptography==2.0.3
daemonize==2.4.7
dnspython==1.15.0
dnspython3==1.15.0
err==3.2.0
errbot==5.1.3
flake8==3.5.0
future==0.16.0
idna==2.6
isort==4.2.15
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
pluggy==0.6.0
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pygments-markdown-lexer==0.1.0.dev39
pylint==1.8.2
pyOpenSSL==17.2.0
pytest==3.3.2
python-telegram-bot==9.0.0
requests==2.18.4
rocket-errbot==1.2.5
salt-pepper==0.5.3
six==1.11.0
typing==3.6.2
urllib3==1.22
virtualenv==15.1.0
waitress==1.1.0
WebOb==1.7.4
WebTest==2.0.29
wrapt==1.10.11
Yapsy==1.11.223
